+++ core/presentation/ui/src/main/java/com/ag_apps/core/presentation/util/BeforeDiscountPriceCalculator.kt
@@ -4,6 +4,6 @@ package com.ag_apps.core.presentation.util
  * @author Ahmed Guedmioui
  */
 
-fun Float.originalPrice(discount: Int): Float {
+fun Double.originalPrice(discount: Int): Double {
     return this * (100 + discount) / 100
 }
